requirements.txt
Brlapi==0.8.2
certifi==2020.6.20
chardet==4.0.0
cupshelpers==1.0
dbus-python==1.2.16
feedparser==5.2.1
idna==2.10
louis==3.16.0
musicbrainzngs==0.7.1
mutagen==1.45.1
numpy==1.19.5
pillow==11.1.0
psycopg2-binary==2.9.10
pycairo==1.16.2
pycups==2.0.1
pygame==1.9.6
PyGObject==3.38.0
pygubu==0.36.3
pyinotify==0.9.6
pysmbc==1.0.23
pyxdg==0.27
quodlibet==4.3.0
requests==2.25.1
six==1.16.0
typing==3.7.4.3
urllib3==1.26.5
xdg==5
